# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WOW Gift Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon?: string;
  description?: string;
}

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  images?: string[];
  categoryId: number;
  category?: Category;
  featured?: boolean;
  customizable?: boolean;
  rating?: number;
  reviewCount?: number;
  sizes?: string[];
  inStock?: boolean;
  createdAt?: string;
}

export interface CreateProductBody {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  images?: string[];
  categoryId: number;
  featured?: boolean;
  customizable?: boolean;
  sizes?: string[];
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  id?: number;
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  customText?: string;
  customImageUrl?: string;
  size?: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  pincode: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  totalAmount: number;
  items?: OrderItem[];
  createdAt?: string;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  price: number;
  customText?: string;
  customImageUrl?: string;
  size?: string;
};

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  pincode: string;
  paymentMethod: CreateOrderBodyPaymentMethod;
  items: CreateOrderBodyItemsItem[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface Review {
  id: number;
  productId: number;
  customerName: string;
  rating: number;
  comment?: string;
  photoUrl?: string;
  createdAt?: string;
}

export interface CreateReviewBody {
  productId: number;
  customerName: string;
  rating: number;
  comment?: string;
  photoUrl?: string;
}

export type DashboardStatsCategoryBreakdownItem = {
  categoryName: string;
  count: number;
};

export interface DashboardStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  categoryBreakdown?: DashboardStatsCategoryBreakdownItem[];
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type ListReviewsParams = {
  productId?: number;
};