# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WOW Gift Shop API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: orders
    description: Order operations
  - name: reviews
    description: Review operations
  - name: dashboard
    description: Dashboard & analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, shipped, delivered, cancelled]
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Review list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Dashboard summary stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Recent orders feed
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/trending products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /categories/{id}/products:
    get:
      operationId: getProductsByCategory
      tags: [categories]
      summary: Get products by category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Products in category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        description:
          type: string
      required: [id, name, slug]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        featured:
          type: boolean
        customizable:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        sizes:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, price, categoryId]

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        featured:
          type: boolean
        customizable:
          type: boolean
        sizes:
          type: array
          items:
            type: string
      required: [name, price, categoryId]

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        pincode:
          type: string
        paymentMethod:
          type: string
          enum: [cod, upi]
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required: [id, customerName, customerPhone, customerAddress, pincode, paymentMethod, status, totalAmount]

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        customText:
          type: string
        customImageUrl:
          type: string
        size:
          type: string
      required: [productId, productName, quantity, price]

    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        pincode:
          type: string
        paymentMethod:
          type: string
          enum: [cod, upi]
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              price:
                type: number
              customText:
                type: string
              customImageUrl:
                type: string
              size:
                type: string
            required: [productId, quantity, price]
      required: [customerName, customerPhone, customerAddress, pincode, paymentMethod, items]

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
      required: [status]

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        photoUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, productId, customerName, rating]

    CreateReviewBody:
      type: object
      properties:
        productId:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        photoUrl:
          type: string
      required: [productId, customerName, rating]

    DashboardStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        categoryBreakdown:
          type: array
          items:
            type: object
            properties:
              categoryName:
                type: string
              count:
                type: integer
            required: [categoryName, count]
      required: [totalProducts, totalOrders, totalRevenue, pendingOrders]