# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WOW Gift Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  pincode: string;
  paymentMethod: CreateOrderBodyPaymentMethod;
  items: CreateOrderBodyItemsItem[];
}