# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WOW Gift Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsCategoryBreakdownItem } from "./dashboardStatsCategoryBreakdownItem";

export interface DashboardStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  categoryBreakdown?: DashboardStatsCategoryBreakdownItem[];
}