# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WOW Gift Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  pincode: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  totalAmount: number;
  items?: OrderItem[];
  createdAt?: Date;
}