# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WOW Gift Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id?: number;
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  customText?: string;
  customImageUrl?: string;
  size?: string;
}