# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WOW Gift Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  images?: string[];
  categoryId: number;
  category?: Category;
  featured?: boolean;
  customizable?: boolean;
  rating?: number;
  reviewCount?: number;
  sizes?: string[];
  inStock?: boolean;
  createdAt?: Date;
}